+++ src/base/b-dynamic-page/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :bug: Bug Fix
+
+* Fixed a bug when the `component` getter is `undefined`. Now the getter can return a promise in that case.
+
 ## v3.0.0-rc.204 (2021-06-23)
 
 #### :bug: Bug Fix